Move metrics initialization out of pkg/metrics
This commit moves InitAllMetrics function from pkg/metrics to a separate
package, pkg/metrics/config. This is done to avoid importing all packages that
define metrics in pkg/metrics. pkg/metrics should act as a library and can be
imported by packages that define metrics, but not the other way round.
